feat: Add comprehensive architecture and UX reviews
This commit introduces a series of detailed architecture and UX reviews of the TeaElephant project. These reviews provide a thorough analysis of the current state of the application and offer a set of actionable recommendations for future improvements.

In addition to the reviews, this commit also includes a number of UI and project structure improvements that were made based on the recommendations in the reviews.